move logging into standalone function
