+++ packages/sdk/src/test/utils/getAnvilLogs.ts
@@ -1,0 +1,10 @@
+export async function getAnvilLogs(url: string, id: number): Promise<string[]> {
+  const response = await fetch(new URL(`${id}/logs`, url), {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+
+  return response.json()
+}
