+++ packages/sdk/src/test/utils/index.ts
@@ -2,6 +2,7 @@ export * from './clients'
 export * from './debug'
 export * from './shellCommand'
 export * from './ethToWei'
+export * from './getAnvilLogs'
 export * from './getWETHBalance'
 export * from './getZora1155Balance'
 export * from './getZora721Balance'
